% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{discography}
\alias{discography}
\title{Fugazi releases data}
\format{
dataframe with one row for each release.
\describe{
\item{releaseid}{numeric id in ascending chronological order, references \code{\link{songs}}}
\item{release}{release name}
\item{releasedate}{release date}
\item{release_date_source}{source of the release date}
}
}
\usage{
discography
}
\description{
Metadata for the Fugazi discography.
}
\examples{
discography
}
\keyword{datasets}
