% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{durations}
\alias{durations}
\title{Fugazi Live Series duration data}
\format{
Data frame with 24513 observations and 4 variables.
\describe{
\item{gid}{show id. References \code{\link{shows}}}
\item{track}{Track number within each album}
\item{song}{Track name. References \code{\link{songs}}}
\item{duration}{Track duration, an hms \code{Period} object}
}
}
\source{
https://www.dischord.com/fugazi_live_series
}
\usage{
durations
}
\description{
Duration data for each Fugazi Live Series audio track,
extracted from MP3 files with \href{https://kid3.kde.org/}{kid3}.
}
\section{Notes}{
 MP3 tags were edited to get a consistent format for each album,
and some corrections were made.
join \code{\link{shows}} on \code{gid} to attach a show's date and other details.
join \code{\link{songs}} on \code{song} to attach details of each song.
}

\examples{
\dontshow{if (requireNamespace("dplyr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# to calculate total duration summed across all tracks. Duration is converted to seconds,
# then summed, then converted back to period format which displays time in normal units.
durations |>
dplyr::group_by() |>
dplyr::summarize(total_duration = lubridate::seconds_to_period(
sum(lubridate::period_to_seconds(duration)))) |>
dplyr::ungroup()
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
