% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{played_with}
\alias{played_with}
\title{Fugazi Live Series data on bands that Fugazi played with
One row per show and band Fugazi played with.
In a few cases extra information was added from other sources.}
\format{
dataframe with one row for each show and band Fugazi played with.
\describe{
\item{gid}{show id, references \code{\link{fls_shows}}}
\item{played_with}{Band name}
}
}
\source{
https://www.dischord.com/fugazi_live_series
}
\usage{
played_with
}
\description{
Fugazi Live Series data on bands that Fugazi played with
One row per show and band Fugazi played with.
In a few cases extra information was added from other sources.
}
\section{Provenance}{
 Derived-cleaned in \code{Repeatr} (its \code{played_with} object). Exported by \code{Repeatr::export_fugazidb_data()}.
}

\examples{
played_with
}
\keyword{datasets}
