% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bands}
\alias{bands}
\title{Fugazi Live Series data on bands that Fugazi played with}
\format{
dataframe with one row for each show and band Fugazi played with.
\describe{
\item{gid}{show id, references \code{\link{shows}}}
\item{band}{Band name}
}
}
\source{
https://www.dischord.com/fugazi_live_series
}
\usage{
bands
}
\description{
One row per show and band Fugazi played with. In a few cases extra
information was added from other sources.
}
\examples{
bands
}
\keyword{datasets}
