% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bands}
\alias{bands}
\title{Fugazi Live Series data on bands that Fugazi played with}
\format{
Data frame with 1669 observations of 2 variables.
\describe{
\item{gid}{Gig id, can be used to join with \code{\link{shows}}}
\item{band}{Band name}
}
}
\source{
https://www.dischord.com/fugazi_live_series
}
\usage{
bands
}
\description{
One row per show and band Fugazi played with. In a few cases extra
information was added from other sources.
}
\examples{
\dontshow{if (requireNamespace("dplyr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# rank the bands by the number of times they played with Fugazi.
bands |>
dplyr::group_by(band) |>
dplyr::summarize(number_shows = dplyr::n()) |>
dplyr::ungroup() |>
dplyr::arrange(desc(number_shows))
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
