% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{fugazi.db-package}
\alias{fugazi.db}
\alias{fugazi.db-package}
\title{fugazi.db: Tidy Data from the Fugazi Live Series}
\description{
tidy data derived from the Fugazi Live Series and related sources -
show listings, venue coordinates, durations, and discographical information.
See \code{vignette("Data-Catalogue", package = "fugazi.db")} for a full
description of every table and the keys used to join them together.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexmitrani/fugazi.db}
  \item Report bugs at \url{https://github.com/alexmitrani/fugazi.db/issues}
}

}
\author{
\strong{Maintainer}: Alex Mitrani \email{27974470+alexmitrani@users.noreply.github.com} (\href{https://orcid.org/0000-0003-1091-7413}{ORCID})

Other contributors:
\itemize{
  \item Dischord Records (copyright holder of the underlying Fugazi Live Series show data - see LICENSE) [copyright holder]
}

}
\keyword{internal}
