% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{fugazibase-package}
\alias{fugazibase}
\alias{fugazibase-package}
\title{fugazibase: Tidy Data from the Fugazi Live Series}
\description{
Tidy reference data documenting the live performance history
of the band Fugazi - show listings, venue coordinates, durations, and
discography - sourced primarily from the Fugazi Live Series
website maintained by Dischord Records. See \code{vignette("Data-Catalogue", package = "fugazibase")}
for a full description of every table and the variables used to join them together.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexmitrani/fugazibase}
  \item Report bugs at \url{https://github.com/alexmitrani/fugazibase/issues}
}

}
\author{
\strong{Maintainer}: Alex Mitrani \email{27974470+alexmitrani@users.noreply.github.com} (\href{https://orcid.org/0000-0003-1091-7413}{ORCID})

Other contributors:
\itemize{
  \item Dischord Records (copyright holder of the Fugazi Live Series data - see LICENSE) [copyright holder]
}

}
\keyword{internal}
