% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{song_tempo_bpm_data}
\alias{song_tempo_bpm_data}
\title{Song tempo BPM data}
\format{
dataframe with one row for each song in the Fugazi discography.
\describe{
\item{song}{The name of the song}
\item{tempo_bpm}{The tempo of the song in beats per minute}
}
}
\source{
Tempos of selected songs measured with the 'liveBPM' app for Android and also finger tapping with a timer.
}
\usage{
song_tempo_bpm_data
}
\description{
Song tempo BPM data
}
\section{Provenance}{
 Raw-hand-curated. Manually compiled per-song tempo.
}

\examples{
song_tempo_bpm_data
}
\keyword{datasets}
