% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{songs}
\alias{songs}
\title{Fugazi studio discography data}
\format{
Data frame with 92 observations of 8 variables.
\describe{
\item{title}{The name of the song - can be used to join this table with \code{\link{durations}}}
\item{rid}{numeric id of the release the song appears on, references \code{\link{discography}}}
\item{release_track}{The song's track number on the studio release (distinct from \code{\link{durations}}'s \code{track}, which numbers a specific live recording)}
\item{instrumental}{Indicates whether or not the piece is an instrumental}
\item{vocals_picciotto}{indicates whether or not Guy Picciotto sang lead vocals on this track}
\item{vocals_mackaye}{indicates whether or not Ian Mackaye sang lead vocals on this track}
\item{vocals_lally}{indicates whether or not Joe Lally sang lead vocals on this track}
\item{release_duration}{The song's duration on its studio release, an hms \code{Period} object (same format as \code{\link{durations}}'s \code{duration})}
}
}
\source{
https://web.archive.org/web/20201112000517/http://en.wikipedia.org/wiki/Fugazi_discography
}
\usage{
songs
}
\description{
One row per song in the Fugazi studio discography, from Wikipedia. The
variables attributing lead vocals are simplifications in some cases where
lead vocals were shared.
}
\examples{
\dontshow{if (requireNamespace("dplyr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# join the discography and songs data frames, order by rid and release_track
discography |>
dplyr::left_join(songs) |>
dplyr::arrange(rid, release_track)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
