% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{songs}
\alias{songs}
\title{Fugazi studio discography data}
\format{
Data frame with 92 observations of 8 variables.
\describe{
\item{title}{The title of the song - can be used to join with \code{\link{durations}}}
\item{rid}{Release id of the album or EP the song appears on, can be used to join with \code{\link{discography}}}
\item{release_track}{The song's track number on the studio release}
\item{instrumental}{Indicates whether or not the piece is an instrumental}
\item{vocals_picciotto}{Indicates whether or not Guy Picciotto sang lead vocals}
\item{vocals_mackaye}{Indicates whether or not Ian Mackaye sang lead vocals}
\item{vocals_lally}{Indicates whether or not Joe Lally sang lead vocals}
\item{release_duration}{The song's duration on the corresponding studio release, an hms \code{Period} object.}
}
}
\source{
https://web.archive.org/web/20201112000517/http://en.wikipedia.org/wiki/Fugazi_discography
}
\usage{
songs
}
\description{
One row per song in the Fugazi studio discography. The
variables attributing lead vocals are simplifications in cases where lead vocals were shared.
}
\examples{
\dontshow{if (requireNamespace("dplyr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# join the discography and songs data frames, order by rid and release_track
discography |>
dplyr::left_join(songs) |>
dplyr::arrange(rid, release_track)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
