% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{songvarslookup}
\alias{songvarslookup}
\title{Fugazi songs data}
\format{
dataframe with one row for each song in the Fugazi discography.
\describe{
\item{releaseid}{numeric id in ascending chronological order}
\item{track_number}{The track number for the song on the release}
\item{song}{The name of the song - the join key used to attach this data to \code{Repeatr}'s show data, not a hardcoded id column, so it can't drift out of sync with song identity assigned there}
\item{instrumental}{Indicates whether or not the piece is an instrumental}
\item{vocals_picciotto}{indicates whether or not Guy Picciotto sang lead vocals on this track}
\item{vocals_mackaye}{indicates whether or not Ian Mackaye sang lead vocals on this track}
\item{vocals_lally}{indicates whether or not Joe Lally sang lead vocals on this track}
\item{duration_seconds}{The duration of the song in seconds}
}
}
\source{
https://web.archive.org/web/20201112000517/http://en.wikipedia.org/wiki/Fugazi_discography
}
\usage{
songvarslookup
}
\description{
Song data from the Fugazi discography pages on Wikipedia. The variables
attributing lead vocals are simplifications in some cases where lead
vocals were shared.
}
\section{Provenance}{
 Raw-hand-curated. Edited by hand against Wikipedia.
}

\examples{
songvarslookup
}
\keyword{datasets}
